==> downsampling_compare/step10_final_results.Rmd <==
---
title: "Final Results"
author: "Noah Klammer"
date: "6/28/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
```

## Clear global env and report

```{r include=FALSE}
rm(list = ls())
gc()
```

# Intro


## Import Rda data

```{r import, message=FALSE, warning=FALSE}
load(file = "hvac_dhw_var_month.Rda")
load(file = "hvac_dhw_var_month_red.Rda")

```

### Set data

```{r}
# CHANGE DATA HERE
df <- hvac_dhw_var_month
df_red <- hvac_dhw_var_month_red
```

# Body

### Sum across rows

```{r}
# maybe summarize verb across rows?

df <- df %>%
  replace(is.na(.), 0) %>%
  mutate(rowsum = rowSums(.)) %>%
  select(rowsum)

df_red <- df_red %>%
  replace(is.na(.), 0) %>%
  mutate(rowsum = rowSums(.)) %>%
  select(rowsum)

month_index <- seq(12)
red_rowsum <- df_red$rowsum
df_com <- cbind(month_index,df,red_rowsum) # comparison df

```


### Definitions of CV-RMSE and NMBE

I am coding working backwards from results. For equations of CV-RMSE and NMBE see, Results chapter.

```{r}
cvrmse <- function(y_i, y_hat_i) {
  sqrt(sum(((y_i - y_hat_i)^2))/(length(y_i)-1))/mean(y_i)
}
nmbe <- function(y_i, y_hat_i) {
  sum(y_i-y_hat_i)/((length(y_i)-1)*mean(y_i))
}
```

# Do stat calcs

```{r}
cvrmse(df_com$rowsum, df_com$red_rowsum)
nmbe(df_com$rowsum, df_com$red_rowsum)
sum(df_com$rowsum)/sum(df_com$red_rowsum)
```

## Plot the residuals

```{r}
plot(df_com$rowsum, df_com$red_rowsum, pch = 20, ylab="Energy use, reduced model (J)", xlab="Energy use, full model (J)", main="Reduced Model versus Full Model")
abline(0,1, col = "red")

```


```{r}
residuals = df_com$rowsum - df_com$red_rowsum

month_index <- seq(12)
#plot(hour_index, residuals, ylab="Residuals", xlab="Year hours", main="Reduced Model versus Full Model") 
#abline(0, 0, col="red")

qqnorm(residuals, frame = FALSE, main = "Normal Q-Q Plot of Residuals")
qqline(residuals, col = "red")


```


# End

<br><br><br>
